feat(rfc-008): P2a — promote mini-jsonschema.mjs to scripts/lib/ (+ shim)
git mv tests/lib/mini-jsonschema.mjs -> scripts/lib/ (engine for the shipped
M1/M2 validator, D2 one-engine); 7-line re-export shim keeps the P0 importer
unchanged (version-hash P1b precedent). Fixes two stale cross-refs in the same
commit (reviewer F-E): the "official-meta-schema validation is owned by P2"
sentence (false under locked D2) and json-instance-validate.mjs's tests/lib
path. test-p0-schemas: 117/0.
